Fixing issues for Coverity - zentf kernel
 - CID-127451: Fixed the Uninitialized scalar variable
 - CID-127269: To avoid Unintentional integer overflow, used
   static_cast to uint64_t
 - CID-127266: Initialized the Uninitialized scalar field
 - CID-127230: Initialized the Uninitialized scalar field
 - CID-127169: Initialized the Uninitialized scalar field
 - CID-127091: Initialized the Uninitialized scalar field
 - CID-127030: Made the changes so that both the comparisons are logical
 - CID-127000: Initialized the Uninitialized scalar field
 - CID-126942: Added a return statement
 - CID-126892: Initialized the Uninitialized scalar field
 - CID-126886: Initialized the Uninitialized scalar field
 - CID-126883: Initialized the Uninitialized scalar field
 - CID-126873: Initialized the Uninitialized scalar field
 - CID-126723: Initialized the Uninitialized scalar field
 - CID-126636: Initialized the Uninitialized scalar field
 - CID-126617: Initialized the Uninitialized pointer field
 - CID-126501: Initialized the Uninitialized scalar field
 - ZENAI-571
 - ZENAI-570
 - Total of 17 issues are fixed in this patch

(cherry picked from commit 84a4918f104e2843eae2a1c229960be1f7a973ce)

+++ tensorflow_plugin/src/amd_cpu/kernels/zendnn/zen_kernel_utils.cc
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2022-2024 Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (c) 2022-2025 Advanced Micro Devices, Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -971,7 +971,8 @@ void ZenConvolution2DBatchNormOrRelu(
     }
     uint64_t bias_offset = 0;
     for (int i = 0; i < batch_size; ++i) {
-      bias_offset = (out_height * out_width) * output_channels * i;
+      bias_offset =
+          static_cast<uint64_t>(out_height) * out_width * output_channels * i;
       zenPostOps(zen_env_obj, static_cast<float *>(output_array),
                  const_cast<const float *>(elementwise_input_new), out_height,
                  out_width, output_channels, output_channels, bias_offset, bias,
